// petforest-fe-ecm/src/utils/enums/reservation.enum.ts
export enum RESERVATION_TIME_STATUS {
  OPENING = 'OPENING',
  CLOSED = 'CLOSED',
  COMPLETED = 'COMPLETED',
}

export enum RESERVATION_TIME_SHIFT {
  AM = 'AM',
  PM = 'PM',
}

export enum SERVICE_TRANSPORT_ACCOMPANIED_STATUS {
  ACCOMPANY = 'ACCOMPANY',
  NOT_ACCOMPANY = 'NOT_ACCOMPANY',
}

export enum VISTOR_NUMBER {
  ONE_PERSON = 'ONE_PERSON',
  TWO_PERSON = 'TWO_PEOPLE',
  THREE_PERSON = 'THREE_PEOPLE',
  FOUR_AND_MORE = 'FOUR_AND_MORE',
}